+++ app/Affiliation.php
@@ -2,12 +2,10 @@
 
 namespace App;
 
-use Illuminate\Database\Eloquent\Model;
+use Awobaz\Compoships\Database\Eloquent\Model;
 
 class Affiliation extends Model
 {
-    use \Awobaz\Compoships\Compoships;
-
     public $primaryKey = ['server', 'node', 'jid'];
     public $incrementing = false;
 }